chore: renamed plugin-inputs.ts to plugin-input.ts
